Update the repo
I returned to this example repo after a long break, and discoverd that
the versions of tools being used were old and that the original layout
of the repo caused the builds to be more complex than they needed to be.

This commit makes frontend and backend more self contained, simplifies
the build process and makes changes required to update tool versions.